+++ src/effects/Paulstretch.h
@@ -50,8 +50,8 @@ private:
    bool ProcessOne(WaveTrack *track, double t0, double t1, int count);
 
 private:
-   float mAmount;
-   float mTime_resolution;  //seconds
+   double mAmount;
+   double mTime_resolution;  //seconds
    double m_t1;
 
    CapturedParameters mParameters;
