+++ src/effects/ScienFilter.cpp
@@ -282,16 +282,16 @@ bool EffectScienFilter::Startup()
       mFilterSubtype = std::min (1, mFilterSubtype);
       gPrefs->Read(base + L"Cutoff", &dTemp, 1000.0);
       mCutoff = (float)dTemp;
-      mCutoff = std::max (1.0f, mCutoff);
-      mCutoff = std::min (1000000.0f, mCutoff);
+      mCutoff = std::max (1.0, mCutoff);
+      mCutoff = std::min (1000000.0, mCutoff);
       gPrefs->Read(base + L"Ripple", &dTemp, 1.0);
       mRipple = dTemp;
-      mRipple = std::max (0.0f, mRipple);
-      mRipple = std::min (100.0f, mRipple);
+      mRipple = std::max (0.0, mRipple);
+      mRipple = std::min (100.0, mRipple);
       gPrefs->Read(base + L"StopbandRipple", &dTemp, 30.0);
       mStopbandRipple = dTemp;
-      mStopbandRipple = std::max (0.0f, mStopbandRipple);
-      mStopbandRipple = std::min (100.0f, mStopbandRipple);
+      mStopbandRipple = std::max (0.0, mStopbandRipple);
+      mStopbandRipple = std::min (100.0, mStopbandRipple);
 
       SaveUserPreset(GetCurrentSettingsGroup());
 
@@ -491,7 +491,7 @@ void EffectScienFilter::PopulateOrExchange(ShuttleGui & S)
          S
             .Id(ID_Ripple)
             .Text(XO("Passband Ripple (dB)"))
-            .Validator<FloatingPointValidator<float>>(
+            .Validator<FloatingPointValidator<double>>(
                1, &mRipple, NumValidatorStyle::DEFAULT,
                Passband.min, Passband.max)
             .Enable( type1Enabler )
@@ -518,7 +518,7 @@ void EffectScienFilter::PopulateOrExchange(ShuttleGui & S)
          S
             .Id(ID_Cutoff)
             .Text(XO("Cutoff (Hz)"))
-            .Validator<FloatingPointValidator<float>>(
+            .Validator<FloatingPointValidator<double>>(
                1, &mCutoff, NumValidatorStyle::DEFAULT,
                Cutoff.min, mNyquist - 1)
             .AddTextBox(XXO("C&utoff:"), L"", 10);
@@ -534,7 +534,7 @@ void EffectScienFilter::PopulateOrExchange(ShuttleGui & S)
          S
             .Id(ID_StopbandRipple)
             .Text(XO("Minimum S&topband Attenuation (dB)"))
-            .Validator<FloatingPointValidator<float>>(
+            .Validator<FloatingPointValidator<double>>(
                1, &mStopbandRipple, NumValidatorStyle::DEFAULT,
                Stopband.min, Stopband.max)
             .Enable( type2Enabler )
@@ -689,7 +689,7 @@ float EffectScienFilter::FilterMagnAtFreq(float Freq)
       break;
 
    case kChebyshevTypeI:     // Chebyshev Type 1
-      double eps; eps = sqrt(pow (10.0, std::max(0.001f, mRipple)/10.0) - 1);
+      double eps; eps = sqrt(pow (10.0, std::max(0.001, mRipple)/10.0) - 1);
       double chebyPolyVal;
       switch (mFilterSubtype)
       {
@@ -706,7 +706,7 @@ float EffectScienFilter::FilterMagnAtFreq(float Freq)
       break;
 
    case kChebyshevTypeII:     // Chebyshev Type 2
-      eps = 1 / sqrt(pow (10.0, std::max(0.001f, mStopbandRipple)/10.0) - 1);
+      eps = 1 / sqrt(pow (10.0, std::max(0.001, mStopbandRipple)/10.0) - 1);
       switch (mFilterSubtype)
       {
       case kLowPass:	// lowpass
